286 Implements the abort process service call
