--- packages/orchestrator-ui-components/src/hooks/ProcessesHooks/useDeleteProcess.ts
@@ -1,37 +1,0 @@
-import { useContext } from 'react';
-import { useMutation, useQueryClient } from 'react-query';
-
-import { signOut } from 'next-auth/react';
-
-import { OrchestratorConfigContext } from '../../contexts';
-import { useSessionWithToken } from '../useSessionWithToken';
-
-export const useDeleteProcess = () => {
-    const { processesEndpoint } = useContext(OrchestratorConfigContext);
-    const { session } = useSessionWithToken();
-    const queryClient = useQueryClient();
-
-    const requestHeaders = {
-        authorization: session ? `Bearer ${session.accessToken}` : '',
-    };
-
-    const deleteProcess = async (id: string) => {
-        const response = await fetch(`${processesEndpoint}/${id}`, {
-            method: 'DELETE',
-            headers: requestHeaders,
-        });
-
-        if (response.status < 200 || response.status >= 300) {
-            console.error(response.status, response.body);
-            if (response.status === 401 || response.status === 403) {
-                signOut();
-            }
-        }
-    };
-
-    return useMutation(deleteProcess, {
-        onSuccess: () => {
-            queryClient.invalidateQueries('processList');
-        },
-    });
-};
